COMPLETE PDF COMPILATION OF THE ASSIGNMENT NOTEBOOK
Unlike the previous PDF document you uploaded which 26 pages, this one is a complete one with all 29 pages. Also, I have added a title and our names at the top of the first page of the PDF document. Kindly use this complete PDF instead of the previous, in case you will upload it for the assignment purpose.